Refactored rendering Reorganized many functions Made rendering functions composable
