--- src/draw/XParticleRenderer.ts
@@ -1,6 +1,0 @@
-import particleRenderer from "./library/particleRenderer.js"
-import text from "./text.js"
-
-export default particleRenderer((location) => {
-  text(() => location, () => "X")
-})
