+++ src/draw/composedRenderingFunctions.ts
@@ -1,21 +1,11 @@
 import { conditional } from "../hof/conditional.js"
 import mapper from "../hof/mapper.js"
-import { isAsteroid, isCargo, isDrone, isType, isOre, isPlayer, isProjectile } from "../hof/conditions.js"
-import { circle } from "./circle.js"
-import playerShipGraphic from "./playerShipGraphic.js"
-import projectileGraphic from "./projectileGraphic.js"
-import Renderer from "./renderer.js"
-import { canvasContextScope } from "./canvasContextScope.js"
+import { isType } from "../hof/conditions.js"
 import compose from "../hof/compose.js"
-import ngon from "./ngon.js"
 import { partial } from "../hof/partial.js"
-import { dotParticleRenderer } from "./dotParticleRenderer.js"
 import passthrough from "../hof/passthrough.js"
-import XParticleRenderer from "./XParticleRenderer.js"
-
-function buildRenderer(condition, draw) {
-  return mapper(conditional(condition, Renderer(canvasContextScope(draw))))
-}
+import { asteroidRenderer, cargoRenderer, droneRenderer, oreRenderer, playerRenderer, projectileRenderer } from "./objectRenderers.js"
+import { dotParticleRenderer, XParticleRenderer } from "./particleRenderers.js"
 
 const particlesToPositionTypeTuples = (time, particles) => particles.map(particle => particle(time))
 export const particleRenderer = time => (particles: Particle[]) => {
@@ -27,12 +17,7 @@ export const particleRenderer = time => (particles: Particle[]) => {
   return renderer(particles)
 }
 
-const asteroidRenderer = buildRenderer(isAsteroid, circle)
-const oreRenderer = buildRenderer(isOre, ngon(4))
-const playerRenderer = buildRenderer(isPlayer, playerShipGraphic)
-const projectileRenderer = buildRenderer(isProjectile, projectileGraphic)
-const droneRenderer = buildRenderer(isDrone, ngon(6))
-const cargoRenderer = buildRenderer(isCargo, ngon(5))
+
 export const gameObjectRenderer = [
   asteroidRenderer,
   playerRenderer,
