--- src/draw/dotParticleRenderer.ts
@@ -1,5 +1,0 @@
-import { unitCircle } from "./circle.js"
-import particleRenderer from "./library/particleRenderer.js"
-
-
-export const dotParticleRenderer = particleRenderer(unitCircle)
