+++ src/draw/graphics/ngon.ts
@@ -1,11 +1,11 @@
-import Vector from "../dataStructures/vector/Vector.js"
-import { context } from "../engine/canvas.js"
-import { flip } from "../hof/flip.js"
-import { partial } from "../hof/partial.js"
-import { rotate } from "./rotate.js"
+import Vector from "../../dataStructures/vector/Vector.js"
+import { context } from "../../engine/canvas.js"
+import { flip } from "../../hof/flip.js"
+import { partial } from "../../hof/partial.js"
+import { rotate } from "../library/rotate.js"
 
 
-const ngon = sides => (location, object) => {
+const ngon = sides => ([location, object]) => {
   const divisionSize = (Math.PI * 2) / sides
   const divisions: number[] = new Array<number>(sides).fill(1).map(
     (_num, index) => (index + 1) * divisionSize
