+++ src/draw/graphics/projectileGraphic.ts
@@ -1,8 +1,8 @@
-import { rotate } from "../draw/rotate.js"
-import { context } from "../engine/canvas.js"
+import { context } from "../../engine/canvas.js"
+import { rotate } from "../library/rotate.js"
 
 //Draws a diamond for the projectile
-const projectileGraphic: RenderFunction<IRotatableRenderable> = (location: TVector, object: IRotatableRenderable) => {
+const projectileGraphic: RenderFunction<IRotatableRenderable> = ([location, object]) => {
   const length = 4
   const width = 2
   const [x, y] = location
