+++ src/draw/library/renderer.ts
@@ -1,6 +1,6 @@
 export default function Renderer(renderfunction: RenderFunction<any>) { //can't coax a type that works here
   return function (object: any) {
-    object.position.forEach(each => renderfunction(each, object))
+    object.position.forEach(each => renderfunction([each, object]))
     //perform the identity so that this function can be mapped and composed
     return object
   }
